I got the raw rect (I'm pretty sure . . ) and the copy rect done.
Copy rect does a double copy - from the framebuffer to a scratch buffer and then from
the scratch buffer back to the framebuffer in the new location.

See note in code.

I am still having a few issues with a tiny blue square every so often. I believe this is happening
during a raw rect update.